last version

+++ brightroute/src/main/java/com/brightroute/brightroute/repository/StudentQuestionsAnswerRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentQuestionsAnswerRepository extends JpaRepository<StudentQuestionsAnswer, Long> {
-    // ممكن تضيف Queries مخصصة هنا لو محتاج
+public interface StudentQuestionsAnswerRepository extends JpaRepository<StudentQuestionsAnswer, Integer> {
+    // CORRECTION: PK type changed from Long to Integer
+    // Custom queries for fetching answers by submission ID would be useful here
 }
