+++ brightroute/src/main/java/com/brightroute/brightroute/repository/StudentQuizSubmissionRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentQuizSubmissionRepository extends JpaRepository<StudentQuizSubmission, Long> {}
+public interface StudentQuizSubmissionRepository extends JpaRepository<StudentQuizSubmission, Integer> {}
